set outstanding pings to zero when a pong is received
The point of pings/pongs is to ensure the server is responding
to messages from the client. Therefore, when a pong is received,
decrementing the number of outstanding pings makes less sense than
setting it to zero as we verified the server is responding to us.
This also matches behaviour in other clients.

Resolves #129