apply connect_timeout to socket creates and tls upgrades
Previously connect_timeout wasn't used when creating a socket or
performing a tls upgrade, this adds it there. It might be better
to apply the timeout to the entire connect process
(create socket, send/receive connect info, tls upgrade) but the
current code structure isn't conducive to that, and I think the
gain is minimal when compared to using any timeout on socket
creation.

Resolves #109, #132