provide a default error callback
If an error callback isn't provided by the caller, an optional
parameter, exceptions are swallowed. By providing a default error
callback if the caller doesn't provide one, exceptions will be
logged so the caller can see them.

Providing a synchronous callback or a coroutine to be executed
as a task are currently deprecated, so this error handling will
not effect those methods.