nats: add typing to main connect function
Currently, the `connect` functions of the main NATS modules are not
fully typed or typed slightly incorrectly. The examples all show that a
single string can be given for the `servers` parameter, but the `Client`
class does not declare that it accepts this type, which causes static
analysis issues when the module is used directly.

This commit extends the signatures for `connect` so that `servers` can
also be a `str`, effectively giving it the same type as what is expected
by `_setup_server_pool`.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>