Fixes to internal flush mechanism plus new connect options
A default pending size of 64 MB like the server is used by default.
This can be changed by using the `pending_size` option on connect,
for example to set it to 8MB only:

```
nats.connect(pending_size=8*1024*1024)
```

And similar to the nats.go client, it can be disabled if set to
be negative:

```
nats.connect(pending_size=-1)
```

When this is done, then any published message during a disconnection
will result in a synchronous `OutboundBufferLimitError` thrown when
publishing.

To control the flushing there is a `flush_timeout` option that can be
passed on connect as well:

```
nats.connect(flush_timeout=10)
```

By default, there is no flush timeout so the client can wait indefinely
during a flush similar to nats.go (eventually the ping interval ought
to unblock the flushing)

Changes:

- Changed `nc.flush()` to now throw `FlushTimeoutError` instead
  which is a subtype of `TimeoutError` that it used be using for backwards
  compatibility.

- Changed EOF disconnections being reported as `StaleConnectionError` and
  they are instead a `UnexpectedEOF` error.

Signed-off-by: Waldemar Quevedo <wally@nats.io>