Modernizes the way nats.py is packaged with PEP-621 [1]
Modern Python packages can contain a pyproject.toml file, first
introduced in PEP 518 and later expanded in PEP 517, PEP 621 and PEP
660 [2].

Starting with PEP 621, the Python community selected pyproject.toml as a
standard way of specifying project metadata [3]. Setuptools has adopted this
standard [4] and will use the information contained in this file as an input
in the build process.

This patch propose o adopt this standard into the nats project to modernize
the way this library is packaged and distribued.

This changes propose to move stored data from setup.py to the
pyproject.toml file by respecting supported keys and sections.

These changes propose to use setuptools as backend to create
distribution packages, because it is a well know backend with a long
story and well maintained and it is the one I'm used to use and so
that's the one I know the well.

Also you should notice that these changes propose to rename our readme
file. All-uppercase letters stand out and make the file easily visible which
makes sense because it is probably the first thing a new user would want
to look at. The file's name is generally written in uppercase [5]. Lowercase
filenames are more common files. Also it will standardize the naming
conventions with others files like `LICENSE`, `CODE_OF_CONDUCT`, etc.

These changes can be tested by using the following commands:
```
$ python -m pip install --upgrade build
$ python -m build
```

[1] https://peps.python.org/pep-0621/
[2] https://packaging.python.org/en/latest/tutorials/packaging-projects/
[3] https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
[4] https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[5] https://en.wikipedia.org/wiki/README