[Fix #325] Removes the server address from reconnection logs in examples.
A couple of examples tries to log the server address during
reconnection, through an inner reconnection callback passed as
parameter to the client object.

However, this kind of callback need to be defined before we instantiate
the client, so the inner callback isn't able to see and access the
client/connection object.

The issue was originaly described into the issue #325, however a couple
of other examples also implement the same logic that surely fails in the
same way (not tested on my side). I only focused my tests on the
example given in the original issue.